fix(bombs): ReimuA orb speed decrement uses the float32 0.4 constant
Native FUN_00407840 (all.c:3656) subtracts _DAT_0048ec74 — a float32
0.4 (0.40000000596…) — times the rate, not the double literal. The 6e-9
constant gap flips the float32 speed store at specific points of the
decel curve (empirically at orb-age 38-39). Does not by itself close
the Phantasm f51879 wall, but removes a real divergence source in the
orb speed stream and protects the speed<-10 state-transition threshold.

Fe25 stays 6/6 (never bombs); Hm54 PASS unaffected (MarisaB).

+++ src/game/player-bombs.ts
@@ -212,12 +212,15 @@ export class BombRunner {
     this.actors.forEach((orb, i) => {
       if (orb.state === 1) {
         // Orb motion is float32 in the exe (FUN_00407840 @ all.c:3656-3706):
-        //   speed = (float)(speed - 0.4*rate)
+        //   speed = (float)(speed - 0.4f*rate)
         //   vx/vy = (float)(cos/sin(angle)*speed)   via FUN_004074e0 (recomputed each frame)
         //   x     = (float)(rate*vx + x)
         // Match the float32 storage so sub-pixel orb positions don't drift — the r=128
         // clear-circle / graze-box boundary otherwise flips a bullet clear-vs-graze.
-        orb.speed = Math.fround(orb.speed - 0.4 * ctx.rate);
+        // The decrement constant _DAT_0048ec74 is a float32 0.4 (0.40000000596…),
+        // not the double literal — the 6e-9 gap flips the f32 speed store at
+        // specific points of the decel curve (observed at orb-age 38-39).
+        orb.speed = Math.fround(orb.speed - Math.fround(0.4) * ctx.rate);
         orb.vx = Math.fround(Math.cos(orb.angle) * orb.speed);
         orb.vy = Math.fround(Math.sin(orb.angle) * orb.speed);
         if (orb.speed < -10) {
